Plug a 32/64 problem on LP64 by using a size_t correctly when doing syscalls.
SVN r5742

+++ runtime/platform/sysctl.c
@@ -12,7 +12,7 @@ size_t GC_pageSize (void) {
 }
 
 size_t GC_totalRam (void) {
-  int physMem;
+  size_t physMem;
   size_t len;
   int mib[2];
